+++ packages/ui-components/web/src/Icons/index.ts
@@ -1,8 +1,24 @@
-// @create-index
-
 export { default as InfoIcon } from './InfoIcon';
+export { default as CloseIcon } from './CloseIcon';
+export { default as MoreIcon } from './MoreIcon';
+export { default as ClearIcon } from './ClearIcon';
+export { default as SearchIcon } from './SearchIcon';
+export { default as DropdownArrowIcon } from './DropdownArrowIcon';
+export { default as CheckIcon } from './CheckIcon';
+export { default as ErrorIcon } from './ErrorIcon';
+export { default as PlusIcon } from './PlusIcon';
+export { default as DownloadIcon } from './DownloadIcon';
+export { default as PauseIcon } from './PauseIcon';
+export { default as PlayIcon } from './PlayIcon';
+export { default as SoundMutedIcon } from './SoundMutedIcon';
+export { default as SoundIcon } from './SoundIcon';
+export { default as MediaReplaceIcon } from './MediaReplaceIcon';
+export { default as ContextMenuIcon } from './ContextMenuIcon';
+export { default as PlaybackIcon } from './PlaybackIcon';
+export { default as ReplaceIcon } from './ReplaceIcon';
+export { default as DeleteIcon } from './DeleteIcon';
 
-// plugin-commons
+//need to import from plugin-commons
 // export { default as AlignCenterIcon } from './AlignCenterIcon';
 // export { default as BackArrowIcon } from './BackArrowIcon';
 // export { default as HeightIcon } from './HeightIcon';
@@ -19,16 +35,8 @@ export { default as InfoIcon } from './InfoIcon';
 // export { default as SpoilerButtonIcon } from './SpoilerButtonIcon';
 // export { default as TrashIcon } from './TrashIcon';
 // export { default as WidthIcon } from './WidthIcon';
-export { default as CloseIcon } from './CloseIcon';
 // export { default as EditIcon } from './EditIcon';
 // export { default as RTLIcon } from './RTLIcon';
 // export { default as LTRIcon } from './LTRIcon';
 // export { default as LoaderIcon } from './LoaderIcon';
 // export { default as HashtagIcon } from './HashtagIcon';
-export { default as MoreIcon } from './MoreIcon';
-export { default as ClearIcon } from './ClearIcon';
-export { default as SearchIcon } from './SearchIcon';
-export { default as DropdownArrowIcon } from './DropdownArrowIcon';
-export { default as CheckIcon } from './CheckIcon';
-export { default as ErrorIcon } from './ErrorIcon';
-export { default as PlusIcon } from './PlusIcon';